fix(core): correct qtm correction factor Q(q,w) calculation
In previous calculation, for small q, the amplitude of R(q,t) is approximate to 1, which after FFT is a strong component at 0 frequency, causing the signal being negligible, so the anser is not right. By removing the DC component of R(q,t), we avoid this mistake.
Additionally, add L2 normalization to Q(q,w) to get the correct amplitude of S(q,w).
Note: for now the amplitude is correct in order of magnitude but miss a pi factor, which should be fixed after finding the reason.